had to move eval around to use in TrainingModule
